Adds a projectRoot option to enable relative cache paths.
Previous to this change cache keys (and the values stored in the cache) were
using absolute paths. This meant that moving a buildable repository to a
different directory would entirely invalidate the cache, and that the cache
wouldn't be shareable between builds in CI when building on different machines.

This change adds a new option `projectRoot` that helps to make the paths
relative, and therefore reusable when the project changes directories.